+++ meetup2apricot/meetup_to_apricot_event_adaptor.py
@@ -113,7 +113,12 @@ class MeetupToApricotEventAdaptor:
                     "AccessControl": {
                         "AccessLevel": self.access_level
                         },
-                    "PaymentMethod": "OnlineOnly"
+                    "PaymentMethod": "OnlineOnly",
+                    "IsWaitlistEnabled": True,
+                    "WaitlistSettings": {
+                        "WaitlistType": "Manual",
+                        "InformationToCollect": "ContactInformation"
+                        }
                     }
                 }